+++ fpy2/ast/syntax_check.py
@@ -2,8 +2,8 @@
 
 from typing import Optional, Self
 
-from ..ast.fpyast import *
-from ..ast.visitor import AstVisitor
+from .fpyast import *
+from .visitor import AstVisitor
 
 from ..utils import FPySyntaxError
